third_party/WebKit/LayoutTests/accessibility/table-caption-expected.txt

This tests that table caption element is exposed for accessibility.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS table.childrenCount is 2
PASS captionText is 'TableCaption'
PASS successfullyParsed is true

TEST COMPLETE
